refactor(__tests__): better file structure of the whole __tests__ file implemented
